refactor: update API routes to use environment variables for authentication endpoints
