+++ src/app/api/user/[id]/route.ts
@@ -1,11 +1,13 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { BASE_URL, API_ENDPOINTS } from '@/lib/constants';
+const BASE_URL = process.env.BASE_URL;
+const USER = process.env.USER;
+
 
 export async function GET(request: NextRequest, { params }: { params: Promise<{ id: string }> }) {
   try {
     const { id } = await params;
 
-    const backendUrl = `${BASE_URL}${API_ENDPOINTS.USERS}/${id}`;
+    const backendUrl = `${BASE_URL}${USER}/${id}`;
 
     const cookieHeader = request.headers.get('cookie') ?? '';
