+++ src/lib/api/api-routes-endpoints.ts
@@ -1,0 +1,6 @@
+export const API_ROUTES = {
+  LOGIN: '/api/auth/login',
+  REGISTER: '/api/auth/register',
+  REFRESH: '/api/auth/refresh',
+  USER_BY_ID: '/api/user',
+} as const;
