+++ src/lib/api/fetch-with-auth.ts
