+++ src/lib/constants.ts
@@ -1,15 +1,7 @@
-
 export const IMAGES = {
   MATES_LOGO: "/mates.png",
 } as const; 
 
 
-export const BASE_URL = "https://localhost:7215";
 
-export const API_ENDPOINTS = {
-  LOGIN: '/api/auth/login',
-  REGISTER: '/api/auth/register',
-  REFRESH: '/api/auth/refresh',
-  USERS: `/users`,
-} as const;
